Bug 1626550 - Update graphite to version 1.3.14. r=lsalzman
Differential Revision: https://phabricator.services.mozilla.com/D69175

--HG--
extra : moz-landing-system : lando

+++ gfx/graphite2/src/Collider.cpp
@@ -1081,8 +1081,8 @@ void SlotCollision::initFromSlot(Segment *seg, Slot *slot)
         return;
     const sparse &p = glyphFace->attrs();
     _flags = p[aCol];
-    _limit = Rect(Position(p[aCol+1], p[aCol+2]),
-                  Position(p[aCol+3], p[aCol+4]));
+    _limit = Rect(Position(int16(p[aCol+1]), int16(p[aCol+2])),
+                  Position(int16(p[aCol+3]), int16(p[aCol+4])));
     _margin = p[aCol+5];
     _marginWt = p[aCol+6];
